requirements.txt
pandas
numpy
scikit-learn
psycopg2-binary
sqlalchemy
python-dotenv
joblib